Create README.md
The application for recording the frequency of all keys pressed on computer keyboards.

This is a C# Windows Forms application to record the counts for all keys pressed on a keyboard. 
The application can help people to understand the behaviour of using keyboards in gaming or typing.
Then it can help people to modify keyboard layout to be more suitable for our customs for using keyboards.
The modification to keyboard layouts can prevent RSI(Repetitive Strain Injury) happened on our fingers.
For example, a finger is very sore after a long time typing,
this application can help you to find out which key you type too many times.

The development tool used is Microsoft Visual Studio 2019.
The used library is User32.dll.